+++ cyclic/cyclic.go
@@ -9,8 +9,8 @@ func NewNumber(max int8) *Number {
 	return &Number{maxValue: max}
 }
 
-func (n *Number) Current() int {
-	return int(n.currentValue)
+func (n *Number) Current() int8 {
+	return n.currentValue
 }
 
 func (n *Number) Reset() {
